Use a consistent logger
